Revert "feat: implement user selection for 'Noted By' field and manage signatures in payroll report"
This reverts commit 12c23f9d327bd309a64aef87d30bd6a4def44959.
