+++ WebClient/src/app/(contextWrapper)/(loggedIn)/(normalUser)/reports/sales/page.tsx
@@ -1,8 +1,10 @@
 "use client";
 
 import { LoadingComponent } from "@/components/LoadingComponent/LoadingComponent";
+import { ReportStatusManager } from "@/components/ReportStatusManager";
 import { SplitButton } from "@/components/SplitButton/SplitButton";
 import * as csclient from "@/lib/api/csclient";
+import type { ReportStatus } from "@/lib/api/csclient/types.gen";
 import { useUser } from "@/lib/providers/user";
 import {
     ActionIcon,
@@ -13,7 +15,7 @@ import {
     Card,
     Flex,
     Group,
-    // Image,
+    Image,
     Modal,
     NumberInput,
     SimpleGrid,
@@ -51,6 +53,21 @@ function SalesandPurchasesContent() {
     const [modalSales, setModalSales] = useState<number>(0);
     const [modalPurchases, setModalPurchases] = useState<number>(0);
     const [deleteModalOpened, setDeleteModalOpened] = useState(false);
+
+    // Signature state management
+    // Reason: Track prepared by (current user) and noted by (selected user) for report signatures
+    const [preparedBy, setPreparedBy] = useState<string | null>(null);
+    const [notedBy, setNotedBy] = useState<string | null>(null); // This will store the user ID, not name
+    const [preparedBySignatureUrl, setPreparedBySignatureUrl] = useState<string | null>(null);
+    const [notedBySignatureUrl, setNotedBySignatureUrl] = useState<string | null>(null);
+
+    // User selection state for "noted by" field
+    const [schoolUsers, setSchoolUsers] = useState<csclient.UserSimple[]>([]);
+    const [selectedNotedByUser, setSelectedNotedByUser] = useState<csclient.UserSimple | null>(null);
+    const [userSelectModalOpened, setUserSelectModalOpened] = useState(false);
+
+    // Report status state
+    const [currentReportStatus, setCurrentReportStatus] = useState<ReportStatus>("draft");
 
     // Fetch entries for the current month
     useEffect(() => {
@@ -91,8 +108,192 @@ function SalesandPurchasesContent() {
         fetchEntries();
     }, [currentMonth, userCtx.userInfo?.schoolId]);
 
+    // Initialize signature data and load school users
+    useEffect(() => {
+        const initializeSignatures = async () => {
+            if (!userCtx.userInfo) return;
+
+            /**
+             * Fetch user signature from the server using their signatureUrn
+             * Reason: Convert stored signature URN to displayable blob URL
+             */
+            const fetchUserSignature = async (signatureUrn: string): Promise<string | null> => {
+                try {
+                    const response = await csclient.getUserSignatureEndpointV1UsersSignatureGet({
+                        query: { fn: signatureUrn },
+                    });
+
+                    // Response data is already a blob, create object URL for display
+                    if (response.data) {
+                        return URL.createObjectURL(response.data as Blob);
+                    }
+                    return null;
+                } catch (error) {
+                    console.error("Failed to fetch user signature:", error);
+                    return null;
+                }
+            };
+
+            /**
+             * Load users from the same school for "noted by" selection
+             * Using the simplified user endpoint to avoid permission errors
+             * Reason: Allow selection of any user from the same school for report approval
+             */
+            const loadSchoolUsers = async () => {
+                if (!userCtx.userInfo?.schoolId) return;
+
+                try {
+                    const response = await csclient.getUsersSimpleEndpointV1UsersSimpleGet();
+
+                    if (response.data) {
+                        // Note: The simple endpoint already filters users to the current user's school
+                        // so we don't need to filter by schoolId here
+                        setSchoolUsers(response.data);
+                    }
+                } catch (error) {
+                    console.error("Failed to load school users:", error);
+                    notifications.show({
+                        title: "Error",
+                        message: "Failed to load users from your school.",
+                        color: "red",
+                    });
+                }
+            };
+
+            /**
+             * Load existing daily report to restore signature information
+             * Reason: If a report was previously submitted with signatures, restore them
+             */
+            const loadExistingReport = async () => {
+                if (!userCtx.userInfo?.schoolId) return;
+
+                try {
+                    const response = await csclient.getSchoolDailyReportV1ReportsDailySchoolIdYearMonthGet({
+                        path: {
+                            school_id: userCtx.userInfo.schoolId,
+                            year: currentMonth.getFullYear(),
+                            month: currentMonth.getMonth() + 1,
+                        },
+                    });
+
+                    if (response.data) {
+                        const report = response.data;
+                        // Load signature information from existing report
+                        // notedBy field contains the user ID, not the name
+                        if (report.notedBy) {
+                            setNotedBy(report.notedBy);
+                        }
+                        // Load report status
+                        if (report.reportStatus) {
+                            setCurrentReportStatus(report.reportStatus as ReportStatus);
+                        }
+                    }
+                } catch {
+                    // If report doesn't exist (404), that's fine - we'll create a new one
+                    console.log("No existing daily report found, starting fresh");
+                }
+            };
+
+            // Set prepared by to current user
+            const currentUserName = `${userCtx.userInfo.nameFirst} ${userCtx.userInfo.nameLast}`.trim();
+            setPreparedBy(currentUserName);
+
+            // Load current user's signature if available
+            if (userCtx.userInfo.signatureUrn) {
+                try {
+                    const signatureUrl = await fetchUserSignature(userCtx.userInfo.signatureUrn);
+                    if (signatureUrl) {
+                        setPreparedBySignatureUrl(signatureUrl);
+                    }
+                } catch (error) {
+                    console.error("Failed to load user signature:", error);
+                }
+            }
+
+            // Load school users and existing report
+            await Promise.all([loadSchoolUsers(), loadExistingReport()]);
+        };
+
+        initializeSignatures();
+    }, [userCtx.userInfo, currentMonth]);
+
+    // Load noted by signature when school users are loaded and notedBy is set
+    useEffect(() => {
+        const loadNotedBySignature = async () => {
+            // Only proceed if we have all necessary data
+            if (!notedBy || !schoolUsers.length || selectedNotedByUser) return;
+
+            // Find the user in schoolUsers that matches the saved notedBy ID
+            const matchingUser = schoolUsers.find((user) => user.id === notedBy);
+
+            if (matchingUser) {
+                setSelectedNotedByUser(matchingUser);
+
+                // Load the user's signature if available
+                if (matchingUser.signatureUrn) {
+                    try {
+                        const response = await csclient.getUserSignatureEndpointV1UsersSignatureGet({
+                            query: { fn: matchingUser.signatureUrn },
+                        });
+
+                        if (response.data) {
+                            const signatureUrl = URL.createObjectURL(response.data as Blob);
+                            setNotedBySignatureUrl(signatureUrl);
+                        }
+                    } catch (error) {
+                        console.error("Failed to load noted by user signature:", error);
+                    }
+                }
+            }
+        };
+
+        loadNotedBySignature();
+    }, [notedBy, schoolUsers, selectedNotedByUser]);
+
     const handleClose = () => {
         router.push("/reports");
+    };
+
+    /**
+     * Handle selection of a user for the "noted by" field
+     * Reason: Load the selected user's signature and update the report data
+     */
+    const handleNotedByUserSelect = async (user: csclient.UserSimple) => {
+        setNotedBy(user.id); // Store the user ID instead of name
+        setSelectedNotedByUser(user);
+
+        // Load the selected user's signature if available
+        if (user.signatureUrn) {
+            try {
+                const response = await csclient.getUserSignatureEndpointV1UsersSignatureGet({
+                    query: { fn: user.signatureUrn },
+                });
+
+                if (response.data) {
+                    const signatureUrl = URL.createObjectURL(response.data as Blob);
+                    setNotedBySignatureUrl(signatureUrl);
+                }
+            } catch (error) {
+                console.error("Failed to load noted by user signature:", error);
+            }
+        } else {
+            setNotedBySignatureUrl(null);
+        }
+
+        setUserSelectModalOpened(false);
+    };
+
+    /**
+     * Clear the noted by user selection
+     * Reason: Allow user to remove the noted by selection and signature
+     */
+    const handleClearNotedBy = () => {
+        setNotedBy(null);
+        setSelectedNotedByUser(null);
+        if (notedBySignatureUrl) {
+            URL.revokeObjectURL(notedBySignatureUrl);
+            setNotedBySignatureUrl(null);
+        }
     };
 
     const handleDateSelect = useCallback(
@@ -314,6 +515,7 @@ function SalesandPurchasesContent() {
             });
             // Find deleted entries (in originalEntries but not in dailyEntries)
             const deletedEntries = originalEntries.filter((e) => !dailyEntries.some((d) => d.day === e.day));
+
             // Bulk create new entries
             if (newEntries.length > 0) {
                 await csclient.createBulkDailySalesAndPurchasesEntriesV1ReportsDailySchoolIdYearMonthEntriesBulkPost({
@@ -355,29 +557,28 @@ function SalesandPurchasesContent() {
                     },
                 });
             }
-            notifications.show({
-                title: "Submission",
-                message: "Your entries have been submitted successfully.",
-                color: "green",
-            });
-            // Refresh data after submit
-            const res = await csclient.getSchoolDailyReportEntriesV1ReportsDailySchoolIdYearMonthEntriesGet({
+
+            // Create/update the daily report with signature information
+            // This ensures the notedBy information is saved with the report
+            await csclient.createSchoolDailyReportV1ReportsDailySchoolIdYearMonthPatch({
                 path: {
                     school_id: userCtx.userInfo.schoolId,
                     year: currentMonth.getFullYear(),
                     month: currentMonth.getMonth() + 1,
                 },
+                query: {
+                    noted_by: notedBy, // Save the noted by information
+                },
             });
-            const entries = (res?.data || []) as csclient.DailyFinancialReportEntry[];
-            const mapped = entries.map((entry) => ({
-                date: entry.parent,
-                day: entry.day,
-                sales: entry.sales,
-                purchases: entry.purchases,
-                netIncome: entry.sales - entry.purchases,
-            }));
-            setDailyEntries(mapped);
-            setOriginalEntries(mapped);
+
+            notifications.show({
+                title: "Success",
+                message: "Your daily sales report has been submitted successfully.",
+                color: "green",
+            });
+
+            // Redirect to reports page after successful submission
+            router.push("/reports");
         } catch (err: unknown) {
             if (err instanceof Error && err.message.includes("404 Not Found")) {
                 return;
@@ -485,15 +686,29 @@ function SalesandPurchasesContent() {
                             </Text>
                         </div>
                     </Group>
-                    <ActionIcon
-                        variant="subtle"
-                        color="gray"
-                        size="lg"
-                        onClick={handleClose}
-                        className="hover:bg-gray-100"
-                    >
-                        <IconX size={20} />
-                    </ActionIcon>
+                    <Group gap="md">
+                        {userCtx.userInfo?.schoolId && (
+                            <ReportStatusManager
+                                currentStatus={currentReportStatus}
+                                reportType="daily"
+                                schoolId={userCtx.userInfo.schoolId}
+                                year={currentMonth.getFullYear()}
+                                month={currentMonth.getMonth() + 1}
+                                onStatusChanged={(newStatus) => {
+                                    setCurrentReportStatus(newStatus);
+                                }}
+                            />
+                        )}
+                        <ActionIcon
+                            variant="subtle"
+                            color="gray"
+                            size="lg"
+                            onClick={handleClose}
+                            className="hover:bg-gray-100"
+                        >
+                            <IconX size={20} />
+                        </ActionIcon>
+                    </Group>
                 </Flex>
 
                 {!userCtx.userInfo?.schoolId && (
@@ -677,40 +892,53 @@ function SalesandPurchasesContent() {
                                     overflow: "hidden",
                                 }}
                             >
-                                {/* <Image src="" alt="Prepared by signature" fit="contain" w="100%" h="100%" /> */}
-                                <Text size="xs" c="dimmed">
-                                    Signature
-                                </Text>
+                                {preparedBySignatureUrl ? (
+                                    <Image
+                                        src={preparedBySignatureUrl}
+                                        alt="Prepared by signature"
+                                        fit="contain"
+                                        w="100%"
+                                        h="100%"
+                                    />
+                                ) : (
+                                    <Text size="xs" c="dimmed">
+                                        Signature
+                                    </Text>
+                                )}
                             </Box>
                             <div style={{ textAlign: "center" }}>
                                 <Text fw={600} size="sm">
-                                    NAME
+                                    {preparedBy || "N/A"}
                                 </Text>
                                 <Text size="xs" c="dimmed">
-                                    Position
+                                    {userCtx.userInfo?.position || "Position"}
                                 </Text>
                             </div>
                         </Stack>
                     </Card>
 
                     {/* Noted By */}
-                    <Card withBorder p="md" style={{ position: "relative" }}>
-                        <Badge
-                            size="sm"
-                            color="orange"
-                            variant="light"
-                            style={{
-                                position: "absolute",
-                                top: "12px",
-                                right: "12px",
-                            }}
-                        >
-                            Status
-                        </Badge>
+                    <Card withBorder p="md">
                         <Stack gap="sm" align="center">
-                            <Text size="sm" c="dimmed" fw={500} style={{ alignSelf: "flex-start" }}>
-                                Noted by
-                            </Text>
+                            <Group justify="space-between" w="100%" align="center">
+                                <Group gap="xs" align="center">
+                                    <Text size="sm" c="dimmed" fw={500}>
+                                        Noted by
+                                    </Text>
+                                    <Badge size="sm" color={selectedNotedByUser ? "green" : "orange"} variant="light">
+                                        {selectedNotedByUser ? "Selected" : "Not Selected"}
+                                    </Badge>
+                                </Group>
+                                {selectedNotedByUser ? (
+                                    <Button size="xs" variant="subtle" color="red" onClick={handleClearNotedBy}>
+                                        Clear
+                                    </Button>
+                                ) : (
+                                    <Button size="xs" variant="light" onClick={() => setUserSelectModalOpened(true)}>
+                                        Select User
+                                    </Button>
+                                )}
+                            </Group>
                             <Box
                                 w={200}
                                 h={80}
@@ -724,22 +952,89 @@ function SalesandPurchasesContent() {
                                     overflow: "hidden",
                                 }}
                             >
-                                {/* <Image src="" alt="Noted by signature" fit="contain" w="100%" h="100%" /> */}
-                                <Text size="xs" c="dimmed">
-                                    Signature
-                                </Text>
+                                {notedBySignatureUrl ? (
+                                    <Image
+                                        src={notedBySignatureUrl}
+                                        alt="Noted by signature"
+                                        fit="contain"
+                                        w="100%"
+                                        h="100%"
+                                    />
+                                ) : (
+                                    <Text size="xs" c="dimmed">
+                                        Signature
+                                    </Text>
+                                )}
                             </Box>
                             <div style={{ textAlign: "center" }}>
                                 <Text fw={600} size="sm">
-                                    NAME
+                                    {selectedNotedByUser
+                                        ? `${selectedNotedByUser.nameFirst} ${selectedNotedByUser.nameLast}`.trim()
+                                        : "N/A"}
                                 </Text>
                                 <Text size="xs" c="dimmed">
-                                    Position
+                                    {selectedNotedByUser?.position || "Position"}
                                 </Text>
                             </div>
                         </Stack>
                     </Card>
                 </SimpleGrid>
+
+                {/* User Selection Modal for "Noted By" */}
+                <Modal
+                    opened={userSelectModalOpened}
+                    onClose={() => setUserSelectModalOpened(false)}
+                    title="Select User for 'Noted By'"
+                    size="md"
+                    centered
+                >
+                    <Stack gap="md">
+                        <Text size="sm" c="dimmed">
+                            Select a user from your school to be noted by on this report:
+                        </Text>
+
+                        {schoolUsers.length === 0 ? (
+                            <Text c="dimmed" ta="center">
+                                No users found from your school.
+                            </Text>
+                        ) : (
+                            <Stack gap="xs">
+                                {schoolUsers.map((user) => (
+                                    <Card
+                                        key={user.id}
+                                        p="sm"
+                                        withBorder
+                                        style={{ cursor: "pointer" }}
+                                        onClick={() => handleNotedByUserSelect(user)}
+                                        className="hover:bg-gray-50"
+                                    >
+                                        <Group justify="space-between">
+                                            <div>
+                                                <Text fw={500}>
+                                                    {user.nameFirst} {user.nameLast}
+                                                </Text>
+                                                <Text size="sm" c="dimmed">
+                                                    {user.position || "No position specified"}
+                                                </Text>
+                                            </div>
+                                            {user.signatureUrn && (
+                                                <Badge size="sm" color="green" variant="light">
+                                                    Has Signature
+                                                </Badge>
+                                            )}
+                                        </Group>
+                                    </Card>
+                                ))}
+                            </Stack>
+                        )}
+
+                        <Group justify="flex-end" mt="md">
+                            <Button variant="outline" onClick={() => setUserSelectModalOpened(false)}>
+                                Cancel
+                            </Button>
+                        </Group>
+                    </Stack>
+                </Modal>
 
                 {/* Edit Modal */}
                 <Modal
